refactor: move group editing modal
