+++ src/actions/modals/edit-group-modal.tsx
@@ -19,10 +19,10 @@ import { getErrorMessage, useFoldersMap, Folder } from '@zextras/carbonio-ui-com
 import { map } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import { MultipleCalendarSelector } from './custom-components/multiple-calendar-selector';
-import { GroupCalendarsList } from './group-calendars-list';
 import { modifyCalendarGroupRequest } from '../../soap/modify-calendar-group-request';
 import { updateCalendarGroupsStore, useGroupById } from '../../store/zustand/calendar-group-store';
+import { MultipleCalendarSelector } from '../../view/sidebar/custom-components/multiple-calendar-selector';
+import { GroupCalendarsList } from '../../view/sidebar/group-calendars-list';
 
 export type EditGroupModalProps = {
 	groupId: string;
